refactor(ai): extract regiment production pin cases under residual fat-pin gate
Streamline the Phase-8 economy regiment build-input production suite into topical case modules and extend the residual fat-pin size gate to keep the thin contract enforced.
